requirements.txt
decord
eva-decord
# for gemini api
# google-genai
gradio
huggingface_hub
imageio
matplotlib
nltk
numpy
omegaconf
openai
opencv-python>=4.4.0.46
openpyxl
pandas
pillow
portalocker
protobuf
python-dotenv
requests
rich
sentencepiece
setuptools
sty
tabulate
tiktoken
timeout-decorator
torch==2.4.1
torchvision
tqdm
transformers
typing_extensions
validators
xlsxwriter
